ДЗ 3

+++ internal/models/progress.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type Progress struct {
-	ID             int        `db:"id" json:"id"`
-	UserID         int        `db:"user_id" json:"user_id"`
-	QuestID        int        `db:"quest_id" json:"quest_id"`
+	ID             int64      `db:"id" json:"id"`
+	UserID         int64      `db:"user_id" json:"user_id"`
+	QuestID        int64      `db:"quest_id" json:"quest_id"`
 	Status         string     `db:"status" json:"status"`
 	CompletedTasks int        `db:"completed_tasks" json:"completed_tasks"`
 	StartedAt      time.Time  `db:"started_at" json:"started_at"`
